Don't freeze value or reason
This has been causing me lots of pain, and it only freezes
shallowly anyway. The problematic case is when the return
of a callback is passed to the next promise. Unless you
make `nil` the last expression of that callback, one of
your precious objects will get frozen, and you'll only
notice when you eventually try to modify it.

Apart from that, immutability is a nice in the Ruby world,
and given that I want promise.rb to play nice with others,
I'm just gonna disregard what the JS-centered A+ spec says.