chore(rake): Stop running flay by default since it is flaky.
It seemed like every change required updating the total_score in its
config, and that total score was also different when running flay with rbx
or jruby.

For some reason running the metrics causes spec to fail on rbx in travis,
so I kept it running just the spec rake task.