Stop using the imperfect reek, flog and flay metrics as part of CI.
These try to enforce rules for code that don't always make sense and
are more often a trade-off that is better decided on by humans. The
symptom of this is the accumulation of exclusions and threshold bumps.
This is especially true when the goals of the code change from
simplicity to performance.